+++ selendroid-large-tests/src/main/java/org/openqa/selendroid/tests/UserRegistrationTest.java
@@ -19,6 +19,7 @@ import org.openqa.selendroid.tests.domain.UserDO;
 import org.openqa.selendroid.tests.internal.BaseAndroidTest;
 import org.openqa.selenium.By;
 import org.openqa.selenium.HasInputDevices;
+import org.openqa.selenium.Point;
 import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
@@ -49,7 +50,10 @@ public class UserRegistrationTest extends BaseAndroidTest {
     WebDriverWait wait = new WebDriverWait(driver, 5);
     WebElement inputUsername =
         wait.until(ExpectedConditions.presenceOfElementLocated(By.id("inputUsername")));
-    CharSequence cs="Selendroid";
+    Point point = inputUsername.getLocation();
+    Assert.assertNotNull(point.x);
+    Assert.assertNotNull(point.y);
+    CharSequence cs = "Selendroid";
 
     ((HasInputDevices) driver).getKeyboard().sendKeys(cs);
     ((HasInputDevices) driver).getKeyboard().sendKeys(SelendroidKeys.MENU);