+++ selendroid-server/src/main/java/org/openqa/selendroid/server/model/SelendroidDriver.java
@@ -47,4 +47,6 @@ public interface SelendroidDriver {
   public String getTitle();
 
   public void get(String url);
+  
+  public TouchScreen getTouch();
 }